Add init action to transitions
If we do a reduction, but we specify that not all of our partition
will be overwritten, LAIK now does the initilization for us.

Add a 2nd way to do the reduction to spmd example: we have a sum
reduction with all tasks having write access to all elements, but
only part of the results will be written.