Fix transition calculation + update spmv2 example
The (previously hanging, outcommented) variant of
spmv2 used a COPY partitioning type, which was not
handled. Ignore partitioning type, as for the
transition actions, only calculated borders are
important.

Default in spmv2 now is to broadcast the partitions
of the input vector directly instead of an all-reduction.
All-reduction still can be requested with "spmv2 -r"