API: no access behavior, but in/out copy requirements
The old API wanted users to specfiy access behavior for
phases of access to LAIK data structures between partitioning
switches. From this, LAIK guessed which data transfers are
needed. Now, we expect the user to explicitly specify
whether data should be preserved/copied from the previous and
to the next phase. While being no function change, this should
make the API much easier to understand.

For example, previously there was a "write" and "write all" behavior,
with the slight difference that the first may only write to some part
of the values, so all values had to be preserved/transferred from
previous phase, in contrast to "write all". It seems much better to
have an explicit API here, specifying which data needs to be copied.

All tests should work