Transition calculation changed for multiple slices
Warning: going from n local to another m local slices
can result in a lot of small "keepLocal" actions.
There also is a nested loop here, which may take long.
TODO: assume sorted slices, and check overlapping of
slices in one sweep.
For now, we assume the number of slices to be small.

Still not working with partitionings with multiple
slices per task, as we do not keep a separate mapping
per slice.