Remove flow flags from partitioning object
We want different LAIK data containers to use the same
partitioning at a given time, but maybe with different
data flow flags. Thus having a flow for a partitioning
makes no sense.

Now the active flow of the access phase a data container
is currently using is stored in the data container.