API: Re-introduce coupling of partitionings
Adds a 4th parameter for creating new partitionings.
If non-null, it specifies a base partitioning, making
the new partitioning one that is coupled to the base.
Whenever the base changes, the partitioners of coupled
partitionings are called to calculate new borders from
the changed borders of the base.

This will be used in a future Jacobi example.