Better mapping reuse
If new mapping would shrink a bit, we keep the larger mapping
and reuse it. Also, reuse mappings not only when we can avoid
copy actions, but also without any need to preserve data.

This is crucial for jac1d, making it faster by more than
a factor of 2, because the essential data was malloc'd/free'd
and memcpy'd all the time...