Default alloc policy: always reuse larger mapping
In spmv2, we switched from all (reduction) to block partitioning,
which resulted in new malloc every time. This is obviously
unneded. Now we keep the all-allocations and temporarily only use
a small part of it.

Instead of a policy, it may be better to have an API call to
enforce reallocation as required (shrinking large mappings).