New tag element for partition slices + 2nd halo partitioner
A partitioner now can specify a tag value (int) for slices
of a partition. If he creates multiple slices with the same
tag (>0), this is a request for LAIK data containers to
put these slices into the same memory mapping (TODO, not in
this commit).

As example, another halo partitioner is provided that does
not include extended corners of original slices (resulting
e.g. in up to 7x the number of needed slices for 3d).
Such a halo partitioner is important for stencils which do
not access diagonal elements (ie. corners), to reduce the
amount of communication.

As the effect on mappings is not yet implemented, we change
the Jacobi examples to use the halo partitioner with the
corners for now, to pass tests.