Avoid eventual buffer overrun in log_flush
We now print line by line to stderr instead of collecting
the lines of a log message to output at once. The idea
was that this avoids mixing lines from different MPI tasks.
However, this was flawed as with huge log messages, one
printf to stderr may result in multiple write syscalls.
With the prefix and sorting, we still get consistent output
even without printing a log message at once. This is used
in tests.

Now the buffered log API should be safe

Still need to convert all logging to use the buffered API