Travis: Use different compilers, not always gcc
I do not really understand the Travis compiler matrix,
but the log showed that "export CC=gcc" is executed when
any gcc version as compiler is requested, and our
'configure' then sets CC for make to 'gcc'.

In Travis build, we just set the CC environment before the
make run, as in
  CC=compiler make
This means that CC as set in Makefile.config (by configure)
is used. To overwrite variables set in Makefile, this has
to be done as argument to make, as in
  make CC=compiler

Thus, our config always use gcc-4.8 (from Trusty), and the
compile matrix was somehow useless :-(

Fix that by explicitely setting CC before running 'configure'.
Also, configure now prints out if a specific compiler is requested.