Introduce key-value store to LAIK Core
This key-value store is expected to be used to synchronize internal LAIK
data among LAIK tasks. It provides a higher abstraction for backends
to implement synchronization than an ad-hoc series of broadcasting of
single values.

The main motivation at the moment is to forward external requests
received by any LAIK task to all other tasks. For this, the KV store
allows value lists in nodes, which are expected to be concatenated
and broadcasted to all at sync time.

It also should be useful to broadcast the results of a partitioner run
only in one task.

TODO: implement sync for MPI backend
Lower priority: add LAIK objects such as spaces/data/... to KV store
to enable consistency checks