Working LAIK markov example + tests
Results are correct, ie. same as with markov-ser.
But hopelessly inefficient at the moment, and only working with
a small number of states (tests use 20 and 40 states).

To work well, this requires
* merging of slices
* merging of transition actions
* merging of data to send/receive
For all of them, currently, a huge number of very small
slices/actions/communications are created.

TODO: add a version compactifying slices, resulting in dense mappings.
For that, "compact" mapping modes has to be supported by LAIK.
Currently, each task maps all states in a sparse way, even though state
values are not needed and invalid, wasting memory.