From Markdown documentation, generate HTML/PDF
Markdown has the benefit to work perfectly with Github,
which automatically converts to nice HTML and working
linking between files.

Using pandoc, this also is easily possible "offline",
both for HTML and PDF (currently using wkhtmltopdf).
This adds a Makefile showing this. For better HTML,
we can add a HTML template, and link Doxygen output.