Introduce flags for partitioners
A partitioning algorithm can provide flags for how
to use its partitioning result. This allows the
partitioner to explicitly ask for slice merging, or
to disable space coverage check.

FIXME: The current slice merging is buggy, therefore
the request for slice merging had to be disabled
in examples/markov.c for now. Otherwise, the test
would fail.