Allow arbitrary large transition objects
Temporary buffers are used, which are enlarged
if required, and then a transition object is
created exactly with the size required.

The idea is that if one transition is large in
an application, other transitions probably also
are large, so no need to free temp buffers.

Now, markov with the pseudo-random connectivity
runs (default with 4 MPI tasks: 8620 slices...).