markov2 example: data flow in other direction as in markov
In markov, we collecting state probabilities from incoming connections,
to update values of owned states, using an overlapping partitioning to
read from.

Instead, in markov2, we propagate the state value owned by a task to other
states using outgoing connections, writing to an overlapping partitioning.
This uses automatic reductions done by LAIK on states at partition
boundaries.

TODO: Still buggy, but not working with >1 tasks anyway at the moment, as
reduction on subsets of an index space is not implemented yet in LAIK.