Fix special-case handling around limited fine-grained reductions
Calculating process groups for fine-grained reductions currently
is limited to 32 processes in input and output group.
This should be enough for real use cases (such as LULESH, which
has a maximum of 8 processes in a fine-grained reduction).

markov2 with a small number of processes sometimes creates
all-to-2 reductions. However, for all-to-X the MPI backend
currently uses MPI_(all)reduce, which fails for all-to-2.
As simple fix, change special case detection such that
all-to-2 is processed as fine-grained reduction. Thus, we
can handle markov2 now with all cases.