Support for multiple mappings with fine-grained reductions
This also fixes a bug in the implementation of fine-grained
reductions: we eventually never sent back results to tasks from
the the output group.

propagation2d does not crash with 1 - 8 tasks (with more, an
assertion in propagation2d.c fails). However, results are
different.. so either a bug with mappings or in propagation2d.