Makefile: Enable additional warnings, but don't yet make warnings fatal.
Since we want this not only for liblaik itself, but also for all other
parts of LAIK, this commit also changes the build system to so that
every Makefile includes the top-level Makefile.common file (which
includes the Makefile.config file). This has the nice side effect of
making the other Makefiles more readable and also streamlines the
compiler options across the project (as far as possible).