API CHANGE: Make indexes and slice borders signed (int64_t)
For abstract index spaces, it makes no sense to force them to be
positive. This is also bad for calculating neighbors at nearside
indexes which easily generates underflows.

However, local indexes and element counts within slices stay.
The reasoning for local indexes is that they always start from
a given base, covering a range of a given number of elements.