Prepare for reservation: MappingList now has pointers to Mapping
With reservations for multiple partitionings in the same LAIK container,
we can have more mappings in total for the container than just for the
active partitioning. Further, we want to maintain a separate MappingList
for each partitioning reserved for. However, some mappings will be the
some for different partitionings. So better use pointers in MappingList.

prepareMaps does the allocation of the Laik_Mapping array still together
with the same allocation as for Laik_MappingList. This needs to be
changed in "reservation" mode.