Add support for srcdir != builddir to configure
Ok, I played a bit around to understand how difficult
this actually would be with plain Makefiles, and
I think that spreading a few "$(SDIR)" into a few
places in the Makefiles is acceptable. Anyway, it does
not hurt for in-place builds.

Now you can do
  cd $LAIKTOP
  mkdir build
  cd build
  ../configure
  make

This also works with "make tests" and "make (un)install".

Configure creates directories and Makefiles mirroring the
source tree, and the generated Makefiles just define
SDIR and include the original Makefile.

This needs slight changes in Makefiles:
- whenever sources are referenced, they must be prefixed
  with $(SDIR) without slash, which is provided in SDIR
- add rule for .%o.%c to include SDIR (the implicit rule
  does not fit anymore)