Fix memory leaks with Laik_Layout structs
Layouts are used in two ways:
(1) to pass into laik_map to specify a layout hint
(2) to specify a concrete layout for a mapping

We mixed up both, and never freed layout structs,
but dynamically allocated them whenever laik_map
was called.

Now, we set "isFixed" to true in a layout to specify whether
a layout is exactly used once, and can be freed afterwards
(a layout may temporarly be referenced twice when re-using
a mapping; freeing reused mappings does not delete the layout.
Otherwise (when "isFixed" is false), the layout object is
shared and must not be freed.

For layout hints, "isFixed" is always set to false.
laik_map() now always uses the same layout hint object.