MacOS: Fix path issue for shared lib by providing abs path
On MacOS, shared libraries should set their "install name",
which is the absolute path. This can be done by providing the
absolute path with -o on linking. This install name then is
stored in binaries linked with that shared library.

This change does not make any difference on Linux, but is
important for linked examples to find liblaik.so on MacOS.
This does not yet work with an installed liblaik.so...