Fix bug in calc_transition for 1d spaces
Currently we do not handle overlapping slices belonging
to same task/process when traversing the 1d space, as we
only remember if a task is interested in one index or not.
So we need to ensure never to see multiple slices active
at one index. To this end, the sorting of slice borders
now ensures that we first "close" a slice before opening
another one.

By luck, on Linux, qsort was a stable sort, which ensured
that the failing situation did not happen, ie. avoiding the
overlap situation.
The bug only became obvious on a Mac with another qsort
implementation... (we should enable MacOS in Travis).