Travis should use the test suite as started from Makfile
We want to use Makefile's and cmake a bit in parallel,
until the cmake port works as expected.

For now, the "reference test suite" to be run by Travis
still should be the one started from the Makefile. We
always can have a "ctest" branch following master and
trigger the test suite from ctest.