Tests: make separate shell script from each test target (1)
This is done in preparation for using the shell scripts also
with CTest. Tests are unchanged (same expected output).

Move all MPI tests into tests/mpi/ directory. The 1-process
MPI run previously was just in tests/, without running it
with mpirun -np 1 ... this only works with OpenMPI.
Now, for tests in tests/, we also test the single backend.