Fix "jac3d -r" + add test cases
The reservation API allocates different mappings
when a partitioner uses different tags, ie. it matches
slices with same tag for reuse of mappings.

Thus, to work with the no-corner halo partitioner (which
generates multiple slices for an original slice, and
uses tags > 0, we also needed to use tag=1 for the bisection
partitioner. This should be a parameter of the partitioner.

TODO: Add test with no-corner halos and 4 tasks (jac3d -n -r)
But there is still a bug: results differ