jac3d-ll: add '-e' for calc/exec of transitions
Pre-calculated transitions still are only on
index spaces, and can be executed for various
data containers (we can do better by going
to low-level actions for data containers).

TODO: transitions should not check complete data
flow from/to, but only needs to care about the flow
between previous phase and the one we transition to.