Combine consecutive GroupReduce actions with same in/out groups
This is not as efficient as combining consecutive
send/recv actions, as there, we always combine all actions
with same peer together.
But GroupReduce actions are sorted by global index, so there
may be multiple ranges of GroupReduce with same in/out, but
interleaved with other actions.

TODO: Use action marking to combine all