Move jac1d to low-level API
Doing a repartitioning between iterations is now
much more explicit, and fixes a memory leak, as
old partitionings are now freed correctly.

When partitionings are used actively in data containers,
we cannot simply run the partitioner again, as the
container then has a wrong idea about data distribution.
Thus, we have to create new partitionings and switch
to them.