Add 3d grid partitioner, use in jac3d with '-g'
With '-g', run a simple loop to find the "best" grid layout
for the given number of processes in <world>. This tries to
minimize the difference between block counts in x/y/z.

The grid partitioner may not assign every process a slice.
When calling laik_map_def1, do not print an error in such cases
but return an empty memory range.