Implement recording of actions in manual reductions
Not used yet, as the recorded action sequence cannot
be executed yet. The MPI backend may not always want
to use manual reductions, but offload to MPI reductions
instead. However, this needs non-blocking collectives
to not result in eventual serialization...

Logging of added actions is implemented.

New classes of actions added for this:
* a "buffer reserve" action with send/recv actions (RBuf)
  referencing the reserve action. Before execution, the
  reservations need to be collected, a real buffer allocated
  and references replaced with pointers into the real buffer.
  Not done yet.
* actions representing callbacks. For call to init/reduce
  functions of LAIK data types.

Furthermore, actions (not all yet) now have a "round" attribute,
which specifies an ordering of execution. We may reorder
send/recv actions in the future, but are not allowed to
mix different rounds. This is important for the manual
reductions: first, input has to be sent to the process
doing the reduction, and then it sends the result to all
interested processes.
For asynchronous execution, the execution of a next round
has to wait for the previous round to finish.