MPI backend: add code to record send/recv ops
First step to get rid of nested loops for correct
order of send/recv for deadlock avoidance.
This will allow to implement different sorting
orders easily by reordering actions, and also do
this reordering with recorded sub-actions of reductions.

Not yet used because sorting for deadlock avoidance not
done yet.