Algorithm to resort action sequence for deadlock avoidance + use it
We now simply append first all sends, then all receives, and
do a final resorting. Should result in same order as before,
but now the resorting is visible in LAIK_LOG=1 output.

This is enabled because we never ask the MPI backend for direct
execution of a transition, but always record the actions (which
do above) before execution.