Compile on Linux: use qsort instead of mergesort
Calling mergesort works on Apple/BSD, but is not
available directly on Linux, so better not use it.
mergesort is available in the BSD-compatibility library
on Linux, but this would need linking with that library,
which is special-casing for Linux.

We can use qsort, as we only sort a pointer list to the
actual actions, thus the addresses of action structs
will give the original order. This can be used to make
qsort stable by using the action addresses in the compare
function.

This is fine as we never will directly sort actions.
The reason is that action structs will have different sizes
soon as this can heavily reduce the space needed for
action sequences.