MPI backend: Fix call to MPI_Reduce + test case
We only are allowed to use MPI_IN_PLACE if we are the root process.

This fixes the MPI error with "mpirun -np 4 markov2 -f 500 5" when
combining reduce actions is enabled. This also is added as test case.

Still, there is a bug in reduction calculation, resulting in wrong output,
when combining reduce is enabled - so, still disabled...