Fix and enable reduce combining in "markov2 -f" test
The issue was that we cannot use input copy ranges, but
need to use separate copy ranges for output of the merged
reduce actions, as output buffers may be different from
input buffers. For markov2, this only happens for the final
reduction in markov2, where result in copied to rank 0
(using a rank0-only all-partition).

Next step is to use our own reductions with multiple primitive
actions, and resort/combine these to get rid of serialization
effects. This finally should fix scalability in LULESH.