tests/tcp/common.sh: Use TCP sockets in the tests.
This makes the tests work on MacOS which doesn't have abstract UNIX sockets.
Unfortunately, this also means that we have to occupy part of the globally
shared TCP port namespace (using a separate networking namespace would require
root priviledges and probably also be Linux-only). This means that the tests
might very well fail because something else has already bound the required
port or because another test is run in parallel.

This commit tries to the reduce the risk of these two things happening by
using the following formula to dynamically allocate ports for the tests:

    port = 10000 + (PID % 5000) * 4 + i       (where i ∈ {0,1,2,3})

This means that the tests will only use ports between 10000 and 30000 which
should hopefully not be used by other services or outgoing TCP connections.
Furthermore, collisions between two tests are only possible if the test
scripts' PIDs are equal (mod 5000). In my tests, this approach has proven to
be very reliable, but it is of course still entirely possible that the tests now
fail because of unrelated network activity or port collisions.

Closes #157.