Allocate mappings after reuse check
Always allocate directly after the reuse check.
Thus, a backend never needs to trigger allocation.

This makes allocation time deterministic: before,
it depended on whether data is written to in a
transition when/whether space for a mapping actually
is allocated.

In the future, we may want to combine actions depending
on whether allocation already happened or not; this
combining/merging must be the same on sender and receiver
side - thus, the sender must be able to assume that an
allocation was done on receiver side.