Fix buffer reservation usage
Appending an action to an action sequence eventually
calls "realloc" if there is not enough space for the
action, thus invalidating any pointers stored to
actions of that sequence.

Thus, do not return a pointer to the BufReserve action
to provoke wrong use, as the pointer easily can get
invalid.
This also fixes 2 wrong usages, which tried to modify
the BufReserve action after adding other actions.