Other send/recv sorting by rank digits
A communication to rank X is in phase p when
the p lower digits of rank number of X is
the same as the p lower digits of my rank.
Within a phase, the lowest non-equal digit
decides who is allowed to send first (the one
with this digit being 0).

This should be better than the simpler sorting
into 2 phases, as we cannot get a serialized
communication as criticial path (to be shown).