Introduce actions to pack/unpack to/from buffer
This also adds implementations to execute this actions
to MPI backend, but this is not MPI-specific and should
be moved out from backend with all other non-MPI-specific
action executions.

TODO: transformation pass that splits PackAndSend/RecvAndUnpack
into separate pack/unpack and send/recv actions, using buffer
reservations large enough to fit complete slices.
This will allow the usage of asynchronous communication, as
the temporary buffers are not reused.