Shorten log prefix, remove phase counter/name
Programmers are lazy, and probably do not set phase info.
So printing this info in every line is a waste. But setting
the new phase will be logged now.

However, the application-set iteration counter is required
to allow for new entering processes to know where to join.
So we programmers will set it if mallability is wanted.