Do any action transformations within same ActionSeq object
Before, we allocated a new Laik_ActionSeq, and copied over
buffers and copy range specifications. This was fragil and
unneeded. Better allow 2 sequences to exist within one
ActionSeq object: one which is active and cannot change,
and one which is currently built/appended. After the build,
we can make the new actions active, deleting the old sequence.

This also reduces allocations, as temporary space for building
the sequence is reused between tranformations, and on activation,
we exactly allocate the space required.