Backend API simplification (1): exec just gets an action sequence
We also remove the wait/probe methods in the backend API for now,
as this will be replaced anyway by a better way for asynchronous
communication (e.g. by executing only parts of the rounds of an
action sequence, with actions in the different rounds being async).

This removes large portions of the MPI backend, which directly
executed a transition on data. Instead, we always do minimal
transformations if we just get a TExec action.
These minimal transformations include the deadlock avoidance
resorting, which previously was done with nested loops in
the direct execution.

We should be able to do the same thing for the TCP backend,
but there are too many changes... :-(