Start reducing memory consumption for action sequences
Currently, we use one struct for every action type,
with a size of 112 bytes (!). Ie. 10 actions >1k.
This made it easy to iterate over instruction, but
waste a lot of space.

Reduce memory consumption by introducing own structs
for each event type, using a common header of 4 bytes,
defined by Laik_Action.
This commit changes all traversals, and introduces
structs for 2 events (TExec and BufReserve).

Now traversal looks like that, with nextAction being a
macro to forward to the next action, using the action size
stored in the header:

  Laik_Action* a = as->action;
  for(int i = 0; i < as->actionCount; i++, a = nextAction(a))
    switch(a->type) {
    case LAIK_AT_XXX: doXXX( (Laik_Action_XXX*) a); break
    ...