Allow switching to a partitioning using another task group
If so, we must be able to migrate the new partitioning to
one with the original group of the old partitioning and back
again after the real switch.
That is, when shrinking, the tasks to remove get assigned
empty partitions before the switch. When expanding, the
added tasks must have assigned empty partitions.