API: remove Laik_AccessPhase from LAIK
Now that the tests do not require access phase API calls anymore,
remove Laik_AccessPhase completely from the LAIK API.

The concept of an access phase was that it allowed to specify the
partitioning to be used by a container in a given program phase
in an indirect way by instead specifying the partitioning algorithm
to run for generating the partitioning on demand when required.
Furthermore, access phases could be configured to use another access
phase as base, ie. using its partitioning as input for a partitioner
run. This allowed to specify a dependency between access phases,
triggering a rerun whenever the base partitioning changed.

While this concept of automatic setting of partitionings by running
the configured partitioner algorithms sounds nice, it makes it
very difficult for LAIK users to reason about when a partitioning
is regenerated. Because partitioner runs may be very time-consuming
and use quite some memory space, it was bad that the access phase
API actually hides away when partitioner runs are triggered.

Instead, it is much better to make it very explicit about when
time/space is consumed, by allowing users to control the execution
of partitioners by generating partitionings directly. Also, when
partitioners use e.g. time measurements for load re-balancing, it
is important for the user to understand when the measurements are
taken and used.

Another problem with the access phase API was that the automatic
switches triggered at arbitrary time required LAIK to know the data
flow at the end of the current access phase, to be able to preserve
data when needed. This needs to be declared by the programmer, and
it results in a confusingly complex API, as users always have
to specify data flows in advance.

Getting rid of access phases now allows to simplify other parts of
the LAIK API.