API: change data flow semantics from in-out to simple flow
Before, a data flow was a tuple of enter/leave semantics, ie.
whether to copy or not to copy data both when entering or leaving
a phase separated by transitions/switches from one partitioning
to another partitioning. Thus, a data flow was a property of
a program phase, and a transition switched from one flow to another.

Now, a data flow is a property of a switch/transition between
partitionings, and it only specifies whether to preserve data
or not. If not, a initialization of the values to the neutral
element of the specified reduction operation can be requested.
This make transitions much simplier and easier to use.

This commit changes all examples/tests. All tests pass.