API: Add space pointer to public Laik_Slice struct
A slice is always defined on a space. Instead of passing multiple
parameters around together with a slice (such as number of dimensions),
better reference the base space in a slice.

This is an API change as Laik_Slice is public.

This commits further adds init functions for Laik_Slice structs
allocated by API users.

Most changes in this commit make sure that the new Laik_Slice::space
member is correctly initialized, including adding a lot of assertions.

Tests work, and most tests are checked to be  Valgrind-clean.

This prepares for some API functions related to slices, as we can get
rid of the "dims" parameter of the used space now. A "dims" parameter
anyway is not adequate for more complex space types introduced in the
future.