Allow to do laik_new_partitioning in separate steps
laik_new_partitioning creates an empty partitioning,
and fills it with slices from a partitioner.

Allow to do this with two function calls, to prepare
for being able to set filters before running the
partitioner, increasing flexibility of what a partitioning
contains,such as filtered views on slices of one task (most
interesting: my own task) or clipping slices to some
index range (for distributed storage of the outcome
of a partitioner algorithm).