Improve logic in 1d transition calculation
For a transition calculation from p1 to p2, we are interested
in any indexes where own process is involved either on sending (input)
or receiving (output) side.

This means that we do not need all slices from p1, but only the ones
which intersect with our own slices from p2, and vice versa.

If we see that e.g. p1 does not have the slices yet we need, we at
least expect own slices from p2 to be able to calculate the intersecting
slices from p1. Some the other way around.