Try fix the intersection stuff for reduced memory consumption
We always should take own slices into account.

Unfortunately, this is not enough: for index ranges a process
cares about, it should see everything other processes do to come
up with the same operations. Just intersecting own with other slices
is not enough for this. Instead, keep intersection with union of
own slices both in from/to partitioning. Not done yet.

Thus, we cannot switch "memory saving in partitionings by doing only
remembering own slices and intersections" on by default yet, as this
breaks tests :-(
It is only used for the "nodes" partitioning in propagation2d for now,
where it works because own slices are the same in from/to partitioning
(it's the same!).