Allow to add 2 intersection filters to a partitioning
To be able to calculate a transition for partititiongs with only own
slices stored, we further need to run the partitioner algorithms again
to get the intersection of own slices in both the from/to partitioning
side. For the latter, 2 intersection filters need to be installed.

Use this in the transition calculation.

However, not yet switched on in general due to some bug...