Allow -1 for task filter: actively set to invalid task
Such a filter always results in no slices stored.

If a process is not part of a group, its rank/id in this
group is set to -1. A partition of an invalid process
in any partitioning is always empty. And thus, a filter
for id -1 results in no slices stored.

Missing/TODO: we should be able to migrate a partitioning
with a task filter set to -1 to another group, to make the
assertion true that ensures that a active task filter uses
the own rank in that group ?! Does not work at the moment.