Simplify task filter: only allow to filter own slices
This makes it easy to migrate a partitioning, as there is no need to change
the filtered task from old to no group. This was buggy before when
a task got removed from a group (changing to "task -1").

FIXME: myfilter not yet activated for all partitionings, as spmv2-shrink test
hangs: a re-run of a partitioner after migration does not create the same
partitioning, as e.g. number of processes changes. However, this is
expected...