Add laik_new_migrated_partitioning()
This creates a partitioning that takes slices from another partitioning
and maps process ranks of slices from the original process group to a new
process group.

This is done by installing a forwarding filter on the other partitioning:
whenever a new slice is appended to the original partitioning, this will
be forwarded also to the other.

This will replace laik_partitioning_migrate(), which is the only function
changing an existing partititioning. It would be good to make partitionings
immutable after filling with slices and freezing, to make the API easier
to understand and use.