Simplify recompile logic on git revision change
Instead of re-generating a C file on revision change,
update git-version.h when revision changes. Only
src/revinfo.c (now commited as regular source) includes
this header and is recompiled when the header
is recreated.

Move the "re-generation on revision change" into a shell
script, as preparation to make this work also with CMake.