Fix CMake build for git revision output
We add a custom target which is always run, and
this calls the script to update git-version.h
on revision change.

An changed git-version.h then triggers a rebuild
of revinfo.o due to the explicitly added object
file dependency of revinfo.c to git-version.h.
CMake docu says that this should not be needed,
but cmake for some reason does not find this
file dependency on its own, even though revinfo.c
includes git-version.h.

Given that the pure Makefile solution for this
"re-compile on revision change" is just a line,
which got extracted to a shell script for this
CMake solution, and using trial and error to
actually get it work, CMake seems much worse than
Makefiles.