Flag for paritionings to only keep own slices
This allows to optionally compile LAIK in a mode
where processes configure all 1d partitionings to only
keep their own slices on a partitioner run, to save memory.
This is enough information to do the allocation of LAIK
containers. However, when a transition calculation needs
to be done, this is not enough and will trigger another
partitioning run, also keeping slices intersecting with
own slices in the source and target partitioning of the
transition.

Note that this does not yet work as expected: tests fail.
Thus the optional compilation using a flag: To enable,
e.g. put "FLAGS+=-DUSE_OWNSLICES" into Makefile.config