Partitioner API change: explicitly provide space + group
The index space and the process group are important parameters
to a partitioner, so it makes sense to provide them explicitly.

More important: the space and group parameters previously
were taken from the partitioning object collecting the slices.
However, it can be useful to run a partitioning algorithm for
a different space and group, and let the partitioning object
do on-the-fly conversion of slices to another space/group at
collection time.