Partitionings should contain all slices if deriving from
E.g. for shrinking in spmv2 with the incremental "reassign" partitioner,
it is not enough if the partitioner knows its own old partition, but
it also needs to know about slices of removed tasks for still existing
ones to take over part of the work.

We simply re-calculate the partitioning if it only kept own slices.