Fix warnings to allow compilation with cmake
These are probably all very buggy fixes, but without a test case,
obviously nobody cares at the moment.

Please make sure in the future to only commit changes which compile
without warnings.

With Makefiles, this commit is not really needed, as in contrast to
the cmake version, it does not treat warnings as errors. However,
I wanted to check that the recent tests which were also added for
cmake actually run.