MPI backend: Also split send in isend/wait
ISend stays at the position of send, and we wait for completion
at end of action sequence.

Also fix a bug in MPI cleanup if an action sequence had no
actions (eventually access to invalid memory).

TODO: check if both send/recv splitting also works with splitting
reductions... should add a test which fails if this is buggy.